Fix issue where userprofile phonenumber not saved
Attempting to save a phone number in the user page form did not save it
to the database, displaying it only from state. Revisiting profile
results in querying an empty node and thus an empty phone number field.
The issue was found in the Mutation resolver where 'args.phone' was
written instead of 'args.phonenumber'. I also added .swp files to
.gitignore (Vim files during editing).