Remove some nested directories and refactor code
This commit addresses changes made in the frontend-work branch and some
in earlier commits. Nested directories hav been removed to simplify
viewing and accessing files.
Directory names changed to lowercase to make it easier to distinguish
between directories and components in paths. Changed instances of '.map'
to '.forEach', as the current use of '.map' does not return any values.
Replaced uses of 'var' to 'let' to ensure proper scoping of variables. Added
extra line spacing to increase code readability.