Add filters to job postings table
Filters do not persist after refresh or going back to the page. Job
postings with type null (not FULLTIME or PARTTIME) will not show in
filtering due to querying limitations. The deadline filter defaults to
in the next 100 days if not specified. Keyword searches are case
sensitive, but I implemented a small function that makes the query
insensitive to the case of the first letter.